Revert "chore: PR feedback - move IntersectionObserver mock to root mocks folder so it can be re-used with other packages. Update import path for the Stage component tests."
This reverts commit 50f05c10698061008fd1c9d463320124f320505e.
