+++ packages/pipeline/src/Stage.spec.tsx
@@ -1,5 +1,4 @@
-// Import mock implementation prior to importing our component to test so that window.IntersectionObserver is mocked out
-import { mockIsIntersecting } from '../../../__mocks__/IntersectionObserver';
+import { mockIsIntersecting } from './mocks/IntersectionObserver';
 
 import React from 'react';
 import { typeIs } from '@leafygreen-ui/lib';
